+++ webapp/src/pages/EssentialInfo/SubPages/ProfileImage.jsx
@@ -1,14 +1,14 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import React, { useRef } from 'react';
 import Button from 'components/Common/Button';
-import useFileUploader from 'hooks/useFileUploader';
+import useFileInput from 'hooks/useFileInput';
 import { useEssentialFormsAction } from 'contexts/EssentialForm/EssentialForm.Provider';
 import * as S from '../EssentialInfo.style';
 
 // TODO: 이미지 업로드 후 디자인
 export default function ProfileImage() {
   const inputFileRef = useRef();
-  const { imageFile, fileHandler, setImageFile } = useFileUploader();
+  const { imageFile, fileHandler, setImageFile } = useFileInput();
   const { onChangeHandler, handleClickNextButton } = useEssentialFormsAction();
 
   const handleOnChange = (event) => {
