+++ webapp/src/pages/SignUp/index.jsx
@@ -49,10 +49,10 @@ export default function SignUp() {
       const response = await authApi.checkDuplicateEmail({ email: inputValues.email });
       const isDuplicated = response.data;
       if (isDuplicated) {
-        notifyNewMessage(notifyDispatch, '이미 사용중인 닉네임입니다!', TOAST_TYPE.Warning);
+        notifyNewMessage(notifyDispatch, '이미 사용중인 이메일니다!', TOAST_TYPE.Warning);
         setIsEmailDuplicate(true);
       } else {
-        notifyNewMessage(notifyDispatch, '사용가능한 닉네임입니다!', TOAST_TYPE.Success);
+        notifyNewMessage(notifyDispatch, '사용가능한 이메일니다!', TOAST_TYPE.Success);
         setIsEmailDuplicate(false);
       }
     } catch (error) {
